% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_logic.R
\name{compare_sentiment}
\alias{compare_sentiment}
\title{Compare Word Vectors}
\usage{
compare_sentiment(doc_vector, sentiment_vector)
}
\arguments{
\item{doc_vector}{A character vector of words from the document.}

\item{sentiment_vector}{A character vector of sentiment keywords (can use *).}
}
\description{
Compares document words to sentiment vectors using loops and wildcards.
}
\examples{
# count <- compare_sentiment(c("nauseated", "happy"), c("nause*", "happ*"))
}
