% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_logic.R
\name{summarize_sentiment}
\alias{summarize_sentiment}
\title{Summarize Sentiment}
\usage{
summarize_sentiment(file_path)
}
\arguments{
\item{file_path}{A string path to the .txt file.}
}
\description{
Summarizes results of positive and negative word matching.
}
\examples{
# results <- summarize_sentiment("diary.txt")
}
